Agregados metodos adicionales, caching y mapstruct a customer-service y account-service

+++ sbms-customer-service/src/main/java/com/bdgarat/sbmscustomerservice/controller/CustomerRESTController.java
@@ -3,9 +3,11 @@ package com.bdgarat.sbmscustomerservice.controller;
 import com.bdgarat.sbmscustomerservice.dto.CustomerDTO;
 import com.bdgarat.sbmscustomerservice.entity.CustomerEntity;
 import com.bdgarat.sbmscustomerservice.exceptions.BadResourceRequestException;
+import com.bdgarat.sbmscustomerservice.exceptions.NoSuchResourceFoundException;
 import com.bdgarat.sbmscustomerservice.service.ICustomerService;
 import lombok.AllArgsConstructor;
 import lombok.extern.log4j.Log4j2;
+import org.apache.coyote.BadRequestException;
 import org.springframework.http.HttpStatus;
 import org.springframework.web.bind.annotation.*;
 import org.springframework.web.server.ResponseStatusException;
@@ -21,63 +23,61 @@ public class CustomerRESTController {
     private ICustomerService customerService;
 
     @GetMapping
-    public List<CustomerDTO> getAllCustomers() {
+    public List<CustomerDTO> getAll() {
         return this.customerService.getAll();
     }
 
     @GetMapping("/cu/{cu}")
-    public CustomerDTO getCustomerByCu(@PathVariable String cu) {
+    public CustomerDTO getByCu(@PathVariable String cu) {
         CustomerDTO customer = this.customerService.getByCu(cu);
         if (customer == null) {
-            log.debug("Customer with Cu {} not found", cu);
-            throw new ResponseStatusException(HttpStatus.NOT_FOUND, "Customer with cu " + cu + " not found");
+            String error = "Customer with Cu " + cu + " not found";
+            log.debug(error);
+            throw new NoSuchResourceFoundException(error);
         }
         return customer;
     }
 
     @GetMapping("/id/{id}")
-    public CustomerDTO getCustomerById(@PathVariable String id) {
+    public CustomerDTO getById(@PathVariable String id) {
         CustomerDTO customer = this.customerService.getById(id);
         if (customer == null) {
-            log.debug("Customer with Id {} not found", id);
-            throw new ResponseStatusException(
-                    HttpStatus.NOT_FOUND, "Customer with id " + id + " not found");
+            String error = "Customer with Id " + id + " not found";
+            log.debug(error);
+            throw new NoSuchResourceFoundException(error);
         }
         return customer;
     }
 
     @PostMapping
     @ResponseStatus(HttpStatus.CREATED)
-    public CustomerDTO addCustomer(@RequestBody CustomerDTO customerDTO) {
-        try {
-            return this.customerService.add(customerDTO);
-        } catch (Exception ex) {
-            log.debug(ex.getMessage(), ex);
-            throw new ResponseStatusException(
-                    HttpStatus.BAD_REQUEST, ex.getMessage(), ex);
+    public CustomerDTO add(@RequestBody CustomerDTO customerDTO) {
+        if(customerDTO == null) {
+            String error = "Customer DTO is null";
+            log.debug(error);
+            throw new BadResourceRequestException(error);
         }
+        return this.customerService.add(customerDTO);
     }
 
     @PutMapping
-    public CustomerDTO updateCustomer(@RequestBody CustomerDTO customerDTO) {
-        try {
-            return this.customerService.update(customerDTO);
-        } catch (Exception ex) {
-            log.debug(ex.getMessage(), ex);
-            throw new ResponseStatusException(
-                    HttpStatus.BAD_REQUEST, ex.getMessage(), ex);
+    public CustomerDTO update(@RequestBody CustomerDTO customerDTO) {
+        if(customerDTO == null) {
+            String error = "Customer DTO is null";
+            log.debug(error);
+            throw new BadResourceRequestException(error);
         }
+        return this.customerService.update(customerDTO);
     }
 
     @DeleteMapping
     @ResponseStatus(HttpStatus.NO_CONTENT)
-    public void deleteCustomer(@RequestBody CustomerDTO customerDTO) {
-        try {
-            this.customerService.delete(customerDTO);
-        } catch (Exception ex) {
-            log.debug(ex.getMessage(), ex);
-            throw new ResponseStatusException(
-                    HttpStatus.BAD_REQUEST, ex.getMessage(), ex);
+    public void delete(@RequestBody CustomerDTO customerDTO) {
+        if(customerDTO == null) {
+            String error = "Customer DTO is null";
+            log.debug(error);
+            throw new BadResourceRequestException(error);
         }
+        this.customerService.delete(customerDTO);
     }
 }
